bugfix in geozero and warning fix in imageMath (#318)
* imageMath: fix numpy deprecation warn

* geozero: bugfix for missing lookSide

+++ applications/imageMath.py
@@ -415,7 +415,7 @@ def main(args, files):
     outBands = iMath['outBands']
 
     ####Array representing columns
-    dataDict['COL'] = np.arange(iMath['width'], dtype=np.float)
+    dataDict['COL'] = np.arange(iMath['width'], dtype=np.float32)
 
     #####Replace ^ by **
     for lineno in range(int(iMath['length'])):
